requirements.txt
fastapi==0.115.12
starlette==0.46.2
uvicorn[standard]==0.34.2
python-multipart==0.0.20
sqlalchemy[asyncio]==2.0.40
aiosqlite==0.21.0
python-jose[cryptography]==3.4.0
bcrypt==4.3.0
pydantic-settings==2.9.1
python-dotenv==1.1.0
httpx[http2]==0.28.1
orjson>=3.9.0
